explanation.txt
I used the command 'touch explanation.txt' to create the file through python.
This command line allows for the user to interact with the computer by ordering it to accomplish the action to create a new file. Thanks to this, we can understand how  the computer stores and organises files.